Fix file_lines issue in examples/mailsettings/mailsettings.java
