+++ examples/mailsettings/UpdateSpamCheckMailSettings.java
@@ -1,0 +1,29 @@
+import java.io.IOException;
+
+import com.sendgrid.Method;
+import com.sendgrid.Request;
+import com.sendgrid.Response;
+import com.sendgrid.SendGrid;
+
+//////////////////////////////////////////////////////////////////
+// Update spam check mail settings
+// PATCH /mail_settings/spam_check
+
+
+public class UpdateSpamCheckMailSettings {
+  public static void main(String[] args) throws IOException {
+    try {
+      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
+      Request request = new Request();
+      request.setMethod(Method.PATCH);
+      request.setEndpoint("mail_settings/spam_check");
+      request.setBody("{\"url\":\"url\",\"max_score\":5,\"enabled\":true}");
+      Response response = sg.api(request);
+      System.out.println(response.getStatusCode());
+      System.out.println(response.getBody());
+      System.out.println(response.getHeaders());
+    } catch (IOException ex) {
+      throw ex;
+    }
+  }
+}
